updated FractalOperators, now for Hutchinson
