+++ src/Fae.jl
@@ -37,8 +37,7 @@ include("utils/histogram.jl")
 include("structs/fractal_input.jl")
 include("structs/fractal_user_methods.jl")
 include("fums/shaders.jl")
-#include("structs/fractal_operators.jl")
-#include("structs/flame_structs.jl")
+include("structs/fractal_operators.jl")
 include("structs/fractal_executable/fractal_executable.jl")
 #include("structs/fractal_executable/hutchinson.jl")
 include("structs/fractal_executable/shader.jl")
